requirements.txt
alembic==1.8.1
autopep8==1.7.0
bcrypt==4.0.0
blinker==1.5
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
packaging==21.3
psycopg2==2.9.3
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.21.0
SQLAlchemy==1.4.41
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
